week7

+++ post/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render
 from rest_framework.response import Response
-from .models import Post
+from .models import Post, Like
 from rest_framework.decorators import api_view
+from tag.models import Tag
 
 @api_view(['POST'])
 def CreatePostView(request):
@@ -30,11 +31,19 @@ class PostListView(APIView):		### 얘네가 class inner function 들! ###
         return Response(serializer.data, status=status.HTTP_200_OK)      
   
     def post(self, request):
+        author = request.user
+        tag_ids = request.data.get('tags')
         title = request.data.get('title')
         content = request.data.get('content')
+        if not author.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
         if not title or not content: # 미리 예상되는 문제에 대해서 분기 처리를 해두는 것이 좋다. 
             return Response({"detail": "[title, content] fields missing."}, status=status.HTTP_400_BAD_REQUEST)
-        post = Post.objects.create(title=title, content=content)
+        for tag_id in tag_ids:
+            if not Tag.objects.filter(id=tag_id).exists():
+                return Response({"detail": "Provided tag not found."}, status=status.HTTP_404_NOT_FOUND)
+        post = Post.objects.create(title=title, content=content, author=author) # author 정보도 담아주쇼
+        post.tags.set(tag_ids)
         serializer = PostSerializer(post)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
@@ -54,6 +63,8 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id)
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -62,6 +73,8 @@ class PostDetailView(APIView):
             post = Post.objects.get(id=post_id) #quertyset
         except:
             return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+        if request.user != post.author:
+            return Response({"detail": "Permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
         title = request.data.get('title')
         content = request.data.get('content')
         if title:
@@ -77,3 +90,25 @@ class PostDetailView(APIView):
         serializer = PostSerializer(post)
         return Response({"Success to Update!"}, status=status.HTTP_200_OK)
     
+class LikeView(APIView):
+    def post(self, request, post_id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials not provided"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        ### 1 ###
+        try:
+            post = Post.objects.get(id=post_id)
+        except:
+            return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+        ### 2 ###
+        like_list = post.like_set.filter(user=request.user)
+
+        ### 3 ###
+        if like_list.count() > 0:
+            post.like_set.get(user=request.user).delete() # 이미 눌렀으면 좋아요 취소
+        else:
+            Like.objects.create(user=request.user, post=post) # 누른 적 없으니까 좋아요 & 좋아요 카운트++
+
+        serializer = PostSerializer(instance=post)
+        return Response(serializer.data, status=status.HTTP_200_OK)
